# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "One Medy API Documentation",
        "description": "One Medy APi Service to be consumed by client Application",
        "contact": {
            "email": "info@walexbiznig.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/drug/prescription/byappointment/": {
            "get": {
                "tags": [
                    "Drug Prescription"
                ],
                "summary": "Get prescription by appointment id",
                "description": "Returns the prescription the specified user",
                "operationId": "getByAppointmentId",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Size"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/drug/prescribe": {
            "post": {
                "tags": [
                    "Drug Prescription"
                ],
                "summary": "doctor creates prescription",
                "operationId": "bb8309bc6a8fd7aeec3a66cfbfc9228a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePrescriptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreatePrescriptionRequest": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "drug_prescription_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "name": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "dosage_description": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "amount": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "total": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "appointment_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/CreatePrescriptionRequest"
                    }
                ]
            },
...
        },
...
    },
...
    "security": [
        []
    ]
}